+++ docs/release-notes.md
@@ -9,6 +9,7 @@ title: jqwik Release Notes
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 ### Table of Contents  
 
+- [1.1.6-SNAPSHOT](#116-snapshot)
 - [1.1.5](#115)
 - [1.1.4](#114)
 - [1.1.3](#113)
@@ -20,6 +21,10 @@ title: jqwik Release Notes
 - [0.8.x](#08x)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
+
+## 1.1.6-SNAPSHOT
+
+- Removed dependencies to internal classes in JUnit platform
 
 ## 1.1.5